requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
coverage==7.8.0
cryptography==44.0.3
ecdsa==0.19.1
fastapi==0.103.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pytest==7.4.0
pytest-cov==4.1.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.23.2
